Separate ini mock files. Corrected box init method and moved to separate util. Removed unused box loot check from spawn ini.
