+++ src/engine/core/database/ini_registry.test.ts
@@ -5,7 +5,6 @@ import {
   DYNAMIC_LTX_PREFIX,
   DYNAMIC_WEATHER_GRAPHS,
   GAME_LTX,
-  PH_BOX_GENERIC_LTX,
   SMART_TERRAIN_MASKS_LTX,
   SOUND_STORIES_LTX,
   SQUAD_BEHAVIOURS_LTX,
@@ -24,7 +23,6 @@ describe("ini_registry database module", () => {
       SYSTEM_INI,
       DUMMY_LTX,
       GAME_LTX,
-      PH_BOX_GENERIC_LTX,
       DYNAMIC_WEATHER_GRAPHS,
       SQUAD_BEHAVIOURS_LTX,
       SMART_TERRAIN_MASKS_LTX,
