--- src/engine/core/managers/box/utils/index.ts
@@ -1,1 +1,0 @@
-export * from "@/engine/core/managers/box/utils/box_utils";
