+++ src/fixtures/xray/mocks/ini/files/default_weather_graphs.ltx.mock.ts
@@ -1,0 +1,19 @@
+export const mockDefaultWeatherGraphs = {
+  dynamic_default: {
+    clear: 0.4,
+    cloudy: 0.4,
+    rain: 0.1,
+    thunder: 0.1,
+  },
+  weather_periods: {
+    neutral: "2, 7",
+    good: "2, 7",
+  },
+  neutral: {
+    cloudy: 0.2,
+    cloudy2: 0.2,
+    cloudy3: 0.2,
+    cloudy4: 0.2,
+    cloudy5: 0.2,
+  },
+};
