+++ src/fixtures/xray/mocks/ini/files/drop_manager.ltx.mock.ts
@@ -1,0 +1,50 @@
+export const mockDropManager = {
+  keep_items: {},
+  item_dependence: {
+    ammo_9x18_fmj: "wpn_pm, wpn_pb, wpn_fort",
+    ammo_9x19_fmj: "wpn_hpsa, wpn_beretta, wpn_walther, wpn_mp5",
+    "ammo_11.43x23_fmj": "wpn_sig220, wpn_colt1911, wpn_usp, wpn_desert_eagle",
+    ammo_12x70_buck: "wpn_bm16, wpn_toz34, wpn_wincheaster1300, wpn_spas12, wpn_protecta",
+  },
+  stalker: {
+    af_cristall: 0,
+    ammo_9x18_fmj: 1,
+    ammo_9x18_pmm: 1,
+    bandage: 0.25,
+    bread: 0.1,
+  },
+  zombied: {
+    af_cristall: 0,
+    ammo_9x18_fmj: 1,
+    medkit: 0.2,
+    medkit_army: 0.05,
+  },
+  item_count_0: {
+    af_cristall: 1,
+    ammo_9x18_fmj: 5,
+  },
+  item_count_1: {
+    af_cristall: 2,
+    ammo_9x18_fmj: 7,
+  },
+  item_count_2: {
+    af_cristall: 3,
+    ammo_9x18_fmj: 10,
+  },
+  item_count_3: {
+    af_cristall: 4,
+    ammo_9x18_fmj: 12,
+  },
+  default: {
+    af_cristall: 10,
+    ammo_9x18_fmj: 11,
+  },
+  zaton: {
+    af_cristall: 21,
+    ammo_9x18_fmj: 22,
+  },
+  pripyat: {
+    af_cristall: 31,
+    ammo_9x18_fmj: 32,
+  },
+};
