+++ src/fixtures/xray/mocks/ini/files/game.ltx.mock.ts
@@ -1,0 +1,5 @@
+export const mockGame = {
+  zaton: {
+    weathers: "dynamic_default",
+  },
+};
