+++ src/fixtures/xray/mocks/ini/files/ph_box_generic.ltx.mock.ts
@@ -1,0 +1,24 @@
+export const mockBoxGeneric = {
+  item_count_0: {},
+  item_count_1: {},
+  item_count_2: {},
+  item_count_3: {},
+  default: {},
+  zaton: {},
+  def_box: {
+    ammo_9x18_fmj: 0,
+    ammo_9x18_pmm: 0,
+    ammo_9x19_pbp: 0,
+  },
+  med_box_low: {
+    bandage: 1,
+    medkit: 1,
+  },
+  med_box_advanced: {
+    bandage: 1,
+    medkit: 0,
+    medkit_army: 1,
+    medkit_scientic: 0,
+    antirad: 1,
+  },
+};
