+++ src/fixtures/xray/mocks/ini/files/simulation_objects_props.ltx.mock.ts
@@ -1,0 +1,7 @@
+export const mockSimulationObjectsProps = {
+  default_squad: {
+    sim_avail: "{+test_squad} true, false",
+    first: "a",
+    second: "b",
+  },
+};
