+++ src/fixtures/xray/mocks/ini/files/spawn.ini.mock.ts
@@ -1,0 +1,5 @@
+export const mockSpawnIni = {
+  story_object: {
+    story_id: "test-story-id",
+  },
+};
