+++ src/fixtures/xray/mocks/ini/files/squad_behaviours.ltx.mock.ts
@@ -1,0 +1,11 @@
+export const mockSquadBehaviours = {
+  stalker: {
+    actor: 1,
+    resource: 2,
+    surge: 1,
+    base: 1,
+    territory: 1,
+    lair: 0,
+    squad: 0,
+  },
+};
