+++ src/fixtures/xray/mocks/ini/files/surge_manager.ltx.mock.ts
@@ -1,0 +1,6 @@
+export const mockSurgeManager = {
+  settings: {
+    condlist: "true",
+    survive: "{+actor_marked_by_zone_3_times} true, {+anabiotic_in_process} true, false",
+  },
+};
