+++ src/fixtures/xray/mocks/ini/files/trade_manager.ltx.mock.ts
@@ -1,0 +1,6 @@
+export const mockTradeManager = {
+  config: {
+    update_period: 3_600_000,
+    resupply_period: 24 * 3_600_000,
+  },
+};
