+++ src/fixtures/xray/mocks/ini/files/travel_manager.ltx.mock.ts
@@ -1,0 +1,23 @@
+export const mockTravelManager = {
+  locations: {
+    zat_sim_1: "st_stalker_zat_sim_1",
+    zat_sim_2: "st_stalker_zat_sim_2",
+    zat_a1: "st_stalker_zat_a1",
+  },
+  traveler: ["zat_stalker_base_smart", "zat_b55", "zat_b100"],
+  zat_stalker_base_smart: {
+    level: "zaton",
+    name: "st_zat_a2_name",
+    condlist: true,
+  },
+  zat_b55: {
+    level: "zaton",
+    name: "st_zat_b55_name",
+    condlist: true,
+  },
+  zat_b100: {
+    level: "zaton",
+    name: "st_zat_b100_name",
+    condlist: true,
+  },
+};
