+++ src/fixtures/xray/mocks/ini/files/treasure_manager.ltx.mock.ts
@@ -1,0 +1,25 @@
+export const mockTreasureManager = {
+  config: {
+    enhanced_mode_enabled: false,
+    rare_cost: 5000,
+    epic_cost: 10000,
+  },
+  list: {
+    jup_b1_secret: null,
+    jup_b2_secret: null,
+    jup_b3_secret: null,
+  },
+  jup_b1_secret: {
+    empty: "{+info_b10_first_zone_visited} true, false",
+    wpn_abakan: "1, 1",
+    wpn_addon_scope: "1, 1",
+  },
+  jup_b2_secret: {
+    refreshing: "true",
+    wpn_abakan: "2, 1",
+  },
+  jup_b3_secret: {
+    wpn_abakan: "1, 1",
+    wpn_addon_scope: "1, 1",
+  },
+};
