+++ src/fixtures/xray/mocks/ini/files/upgrades_manager.ltx.mock.ts
@@ -1,0 +1,5 @@
+export const mockUpgradesManager = {
+  config: {
+    item_repair_price_coefficient: 0.6,
+  },
+};
